+++ lib/features/appointments/open_appointment_panel.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:apexo/app/routes.dart';
+import 'package:apexo/common_widgets/dental_chart.dart';
 import 'package:apexo/common_widgets/dialogs/import_photos_dialog.dart';
 import 'package:apexo/features/patients/patient_model.dart';
 import 'package:apexo/utils/imgs.dart';
@@ -31,7 +32,9 @@ void openAppointment([Appointment? appointment]) {
     item: editingCopy,
     store: appointments,
     icon: FluentIcons.calendar,
-    title: appointments.get(editingCopy.id) == null ? txt("addAppointment") : editingCopy.title,
+    title: appointments.get(editingCopy.id) == null
+        ? txt("addAppointment")
+        : editingCopy.title,
     tabs: [],
   );
 
@@ -51,7 +54,10 @@ void openAppointment([Appointment? appointment]) {
       icon: FluentIcons.camera,
       body: _AppointmentGallery(panel),
       onlyIfSaved: true,
-      footer: kIsWeb ? null : _AppointmentGalleryFooter(panel), // TODO: image upload isn't supported on web
+      footer: kIsWeb
+          ? null
+          : _AppointmentGalleryFooter(
+              panel), // TODO: image upload isn't supported on web
       padding: 0,
     ),
   ];
@@ -64,7 +70,8 @@ class _AppointmentGalleryFooter extends StatefulWidget {
   const _AppointmentGalleryFooter(this.panel);
 
   @override
-  State<_AppointmentGalleryFooter> createState() => _AppointmentGalleryFooterState();
+  State<_AppointmentGalleryFooter> createState() =>
+      _AppointmentGalleryFooterState();
 }
 
 class _AppointmentGalleryFooterState extends State<_AppointmentGalleryFooter> {
@@ -103,15 +110,18 @@ class _AppointmentGalleryFooterState extends State<_AppointmentGalleryFooter> {
                   ],
                 ),
                 onPressed: () async {
-                  final XFile? res = await ImagePicker().pickImage(source: ImageSource.camera);
+                  final XFile? res =
+                      await ImagePicker().pickImage(source: ImageSource.camera);
                   if (res == null) return;
                   widget.panel.inProgress(true);
                   try {
-                    final imgName = await handleNewImage(rowID: widget.panel.item.id, targetPath: res.path);
+                    final imgName = await handleNewImage(
+                        rowID: widget.panel.item.id, targetPath: res.path);
                     if (widget.panel.item.imgs.contains(imgName) == false) {
                       widget.panel.item.imgs.add(imgName);
                       appointments.set(widget.panel.item);
-                      widget.panel.savedJson = jsonEncode(widget.panel.item.toJson());
+                      widget.panel.savedJson =
+                          jsonEncode(widget.panel.item.toJson());
                     }
                   } catch (e, s) {
                     logger("Error during uploading camera capture: $e", s);
@@ -129,15 +139,18 @@ class _AppointmentGalleryFooterState extends State<_AppointmentGalleryFooter> {
                 ],
               ),
               onPressed: () async {
-                List<XFile> res = await ImagePicker().pickMultiImage(limit: 50 - widget.panel.item.imgs.length);
+                List<XFile> res = await ImagePicker()
+                    .pickMultiImage(limit: 50 - widget.panel.item.imgs.length);
                 widget.panel.inProgress(true);
                 try {
                   for (var img in res) {
-                    final imgName = await handleNewImage(rowID: widget.panel.item.id, targetPath: img.path);
+                    final imgName = await handleNewImage(
+                        rowID: widget.panel.item.id, targetPath: img.path);
                     if (widget.panel.item.imgs.contains(imgName) == false) {
                       widget.panel.item.imgs.add(imgName);
                       appointments.set(widget.panel.item);
-                      widget.panel.savedJson = jsonEncode(widget.panel.item.toJson());
+                      widget.panel.savedJson =
+                          jsonEncode(widget.panel.item.toJson());
                       widget.panel.selectedTab(widget.panel.selectedTab());
                     }
                   }
@@ -170,7 +183,10 @@ class _AppointmentGalleryState extends State<_AppointmentGallery> {
         stream: widget.panel.selectedTab.stream,
         builder: (context, _) {
           return widget.panel.item.imgs.isEmpty
-              ? Center(child: InfoBar(title: Txt(txt("emptyGallery")), content: Txt(txt("noPhotos"))))
+              ? Center(
+                  child: InfoBar(
+                      title: Txt(txt("emptyGallery")),
+                      content: Txt(txt("noPhotos"))))
               : StreamBuilder(
                   stream: widget.panel.inProgress.stream,
                   builder: (context, snapshot) {
@@ -181,10 +197,12 @@ class _AppointmentGalleryState extends State<_AppointmentGallery> {
                       onPressDelete: (img) async {
                         widget.panel.inProgress(true);
                         try {
-                          await appointments.deleteImg(widget.panel.item.id, img);
+                          await appointments.deleteImg(
+                              widget.panel.item.id, img);
                           widget.panel.item.imgs.remove(img);
                           appointments.set(widget.panel.item);
-                          widget.panel.savedJson = jsonEncode(widget.panel.item.toJson());
+                          widget.panel.savedJson =
+                              jsonEncode(widget.panel.item.toJson());
                         } catch (e, s) {
                           logger("Error during deleting image: $e", s);
                         }
@@ -215,28 +233,32 @@ class _AppointmentDetailsState extends State<_AppointmentDetails> {
           /// rebuild needed if a patient is selected/deselected
           key: Key(widget.appointment.patientID ?? ""),
           label: "${txt("patient")}:",
-          child: Row(mainAxisSize: MainAxisSize.min, mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
-            Expanded(
-              child: PatientPicker(
-                  value: widget.appointment.patientID,
-                  onChanged: (id) {
-                    setState(() {
-                      widget.appointment.patientID = id;
-                    });
-                  }),
-            ),
-            const SizedBox(width: 5),
-            if (widget.appointment.patientID == null)
-              AcrylicButton(
-                  icon: FluentIcons.add_friend,
-                  text: txt("newPatient"),
-                  onPressed: () async {
-                    final newPatientId = uuid();
-                    final newPatient = await openPatient(Patient.fromJson({"id": newPatientId}));
-                    routes.closePanel(newPatientId);
-                    widget.appointment.patientID = newPatient.id;
-                  })
-          ]),
+          child: Row(
+              mainAxisSize: MainAxisSize.min,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Expanded(
+                  child: PatientPicker(
+                      value: widget.appointment.patientID,
+                      onChanged: (id) {
+                        setState(() {
+                          widget.appointment.patientID = id;
+                        });
+                      }),
+                ),
+                const SizedBox(width: 5),
+                if (widget.appointment.patientID == null)
+                  AcrylicButton(
+                      icon: FluentIcons.add_friend,
+                      text: txt("newPatient"),
+                      onPressed: () async {
+                        final newPatientId = uuid();
+                        final newPatient = await openPatient(
+                            Patient.fromJson({"id": newPatientId}));
+                        routes.closePanel(newPatientId);
+                        widget.appointment.patientID = newPatient.id;
+                      })
+              ]),
         ),
         InfoLabel(
           label: "${txt("doctors")}:",
@@ -269,7 +291,8 @@ class _AppointmentDetailsState extends State<_AppointmentDetails> {
             ),
             const SizedBox(height: 5),
             if (widget.appointment.operators.isNotEmpty &&
-                !widget.appointment.availableWeekDays.contains(widget.appointment.date.weekday))
+                !widget.appointment.availableWeekDays
+                    .contains(widget.appointment.date.weekday))
               InfoBar(
                 title: Txt(txt("attention")),
                 content: Txt(txt("doctorNotAvailable")),
@@ -303,7 +326,8 @@ class _AppointmentDetailsState extends State<_AppointmentDetails> {
             key: WK.fieldAppointmentPreOpNotes,
             expands: true,
             maxLines: null,
-            controller: TextEditingController(text: widget.appointment.preOpNotes),
+            controller:
+                TextEditingController(text: widget.appointment.preOpNotes),
             onChanged: (v) => widget.appointment.preOpNotes = v,
             placeholder: "${txt("preOperativeNotes")}...",
           ),
@@ -363,18 +387,34 @@ class _OperativeDetailsState extends State<_OperativeDetails> {
             placeholder: "${txt("postOperativeNotes")}...",
           ),
         ),
+        Expander(
+          headerBackgroundColor: WidgetStatePropertyAll(Colors.grey.withAlpha(15)),
+          leading: const Icon(FluentIcons.teeth),
+          header: InfoLabel(label: "${txt("dentalNotes")} (${widget.appointment.teeth.length}):"),
+          content: DentalChart(
+            patient: widget.appointment.patient!,
+            appointment: widget.appointment,
+          ),
+        ),
         InfoLabel(
           label: "${txt("prescription")}:",
           child: TagInputWidget(
             key: WK.fieldAppointmentPrescriptions,
-            suggestions: appointments.allPrescriptions.map((p) => TagInputItem(value: p, label: p)).toList(),
+            suggestions: appointments.allPrescriptions
+                .map((p) => TagInputItem(value: p, label: p))
+                .toList(),
             onChanged: (s) {
               setState(() {
-                widget.appointment.prescriptions = s.where((x) => x.value != null).map((x) => x.value!).toList();
+                widget.appointment.prescriptions = s
+                    .where((x) => x.value != null)
+                    .map((x) => x.value!)
+                    .toList();
                 widget.appointment.isDone = true;
               });
             },
-            initialValue: widget.appointment.prescriptions.map((p) => TagInputItem(value: p, label: p)).toList(),
+            initialValue: widget.appointment.prescriptions
+                .map((p) => TagInputItem(value: p, label: p))
+                .toList(),
             strict: false,
             limit: 999,
             placeholder: "${txt("prescription")}...",
@@ -385,7 +425,11 @@ class _OperativeDetailsState extends State<_OperativeDetails> {
               style: const ButtonStyle(elevation: WidgetStatePropertyAll(2)),
               child: Row(
                 mainAxisSize: MainAxisSize.min,
-                children: [const Icon(FluentIcons.print), const SizedBox(width: 10), Txt(txt("printPrescription"))],
+                children: [
+                  const Icon(FluentIcons.print),
+                  const SizedBox(width: 10),
+                  Txt(txt("printPrescription"))
+                ],
               ),
               onPressed: () {
                 printingPrescription(
@@ -401,7 +445,8 @@ class _OperativeDetailsState extends State<_OperativeDetails> {
           children: [
             Expanded(
               child: InfoLabel(
-                label: "${txt("priceIn")} ${globalSettings.get("currency_______").value}",
+                label:
+                    "${txt("priceIn")} ${globalSettings.get("currency_______").value}",
                 child: CupertinoTextField(
                   key: WK.fieldAppointmentPrice,
                   controller: priceController,
@@ -410,7 +455,8 @@ class _OperativeDetailsState extends State<_OperativeDetails> {
                       widget.appointment.price = double.tryParse(v) ?? 0;
                       if (didNotEditPaidYet) {
                         widget.appointment.paid = double.tryParse(v) ?? 0;
-                        paidController.text = widget.appointment.paid.toStringAsFixed(0);
+                        paidController.text =
+                            widget.appointment.paid.toStringAsFixed(0);
                       }
                       widget.appointment.isDone = true;
                     });
@@ -424,7 +470,8 @@ class _OperativeDetailsState extends State<_OperativeDetails> {
             const SizedBox(width: 10),
             Expanded(
               child: InfoLabel(
-                label: "${txt("paidIn")} ${globalSettings.get("currency_______").value}",
+                label:
+                    "${txt("paidIn")} ${globalSettings.get("currency_______").value}",
                 child: CupertinoTextField(
                   key: WK.fieldAppointmentPayment,
                   controller: paidController,